5-1-16-r6

+++ scripts/check.sh
@@ -5,16 +5,27 @@ function show_all_processes {
     cols=$(tput cols)
     export COLUMNS=1000
 
-    ps -Af |\
-    grep -E 'coffee|compass|sass' |\
-    grep -v grep |\
-    awk '{print substr($2" "substr($0, index($0,$8)), 0, '$cols')}'
+    out=$(
+        ps -Af |\
+        grep -E 'coffee|compass|sass' |\
+        grep -v grep
+    )
+
+    if [ "$1" = "full" ]; then
+        echo "$out" | awk '{print $2" "substr($0, index($0,$8))}'
+    else
+        echo "$out" | awk '{print substr($2" "substr($0, index($0,$8)), 0, '$cols')}'
+    fi
 }
 
 function kill_all_processes {
     out_show=$(show_all_processes)
-    echo -en "$out_show"
-    kill $(echo -en "$out_show" | awk '{print $1}')
+    if [ "$out_show" ]; then
+        echo -en "$out_show"
+        kill $(echo -en "$out_show" | awk '{print $1}')
+    else
+        echo 'none'
+    fi
 }
 
 
@@ -24,6 +35,9 @@ case $1 in
         ;;
     watch)
         watch -n1 $0
+        ;;
+    full)
+        show_all_processes full
         ;;
     *)
         show_all_processes
